refactor(agent-engine): separate turn input from transitions
