+++ crates/agent-engine/src/runtime/mod.rs
@@ -35,8 +35,8 @@ mod tool_policy;
 mod tool_presentation;
 mod tool_resolution;
 mod transition;
-mod turn_driver;
 mod turn_executor;
+mod turn_input;
 mod turn_memory;
 mod turn_support;
 mod ui_surfaces;
