+++ crates/agent-engine/src/runtime/tool_batch.rs
@@ -2,7 +2,7 @@
 
 use tokio_util::sync::CancellationToken;
 
-use super::turn_driver::TurnInputBroker;
+use super::turn_input::TurnInputBroker;
 use crate::agent_machine::NormalizedToolCall;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
