+++ crates/agent-engine/src/runtime/turn_executor.rs
@@ -13,7 +13,7 @@ use super::response_guardrails::{
 };
 use super::tool_batch::{ToolBatchOrchestratorOutcome, ToolOrchestratorInputs};
 use super::transition::{RuntimeLoopState, orchestrate_tool_batch};
-use super::turn_driver::TurnInputBroker;
+use super::turn_input::TurnInputBroker;
 use super::turn_memory::{FinalizeTurnMemoryRequest, finalize_turn_memory_best_effort};
 use super::turn_support::{
     check_turn_cancelled, emit_streaming_chunks, emit_task_completed_success, emit_thinking_chunks,
